change translateMonthToSpanish to utils folder
