+++ lib/domain/entities/calculate_investment.dart
@@ -1,4 +1,5 @@
 import 'package:finniu/domain/entities/plan_entities.dart';
+import 'package:finniu/utils/translate_month_to_spanish.dart';
 
 class PlanSimulation {
   final int initialAmount;
@@ -36,27 +37,4 @@ class PlanSimulation {
       return 'N/A';
     }
   }
-}
-
-String translateMonthToSpanish(String date) {
-  Map<String, String> months = {
-    'January': 'Enero',
-    'February': 'Febrero',
-    'March': 'Marzo',
-    'April': 'Abril',
-    'May': 'Mayo',
-    'June': 'Junio',
-    'July': 'Julio',
-    'August': 'Agosto',
-    'September': 'Septiembre',
-    'October': 'Octubre',
-    'November': 'Noviembre',
-    'December': 'Diciembre',
-  };
-
-  months.forEach((key, value) {
-    date = date.replaceAll(key, value);
-  });
-
-  return date;
 }
