+++ lib/utils/translate_month_to_spanish.dart
@@ -1,0 +1,22 @@
+String translateMonthToSpanish(String date) {
+  Map<String, String> months = {
+    'January': 'Enero',
+    'February': 'Febrero',
+    'March': 'Marzo',
+    'April': 'Abril',
+    'May': 'Mayo',
+    'June': 'Junio',
+    'July': 'Julio',
+    'August': 'Agosto',
+    'September': 'Septiembre',
+    'October': 'Octubre',
+    'November': 'Noviembre',
+    'December': 'Diciembre',
+  };
+
+  months.forEach((key, value) {
+    date = date.replaceAll(key, value);
+  });
+
+  return date;
+}
